\name{demographics}
\alias{demographics}
\docType{data}
\title{
demographics/data.frame
}
\description{
dataframe of 50 records and with columns id, age, gender,sbp, dbp, smoker
}
\usage{data(demographics)}
\format{
  The format is:
 chr "demographics"
}
\examples{
data(demographics)
## maybe str(demographics) ; plot(demographics) ...
}
\keyword{datasets}
