\name{genotype}
\alias{genotype}
\docType{data}
\title{
genotype/data.frame
}
\description{
dataframe of 459 records and with character columns id, snp_id, genotype
}
\usage{data(genotype)}
\format{
  The format is:
 chr "genotype"
}
\examples{
data(genotype)
## maybe str(genotype) ; plot(genotype) ...
}
\keyword{datasets}
