\name{snp}
\alias{snp}
\docType{data}
\title{
snp/data.frame
}
\description{
dataframe of 10 records and with character columns snp_id, chr, all_A, all_B
}
\usage{data(snp)}
\format{
  The format is:
 chr "snp"
}
\details{
test dataset
}
\examples{
data(snp)
## maybe str(snp) ; plot(snp) ...
}
\keyword{datasets}
